+++ backend/src/api/endpoints/memo.py
@@ -18,11 +18,11 @@ from app.core.data.dto.memo import (
 )
 from app.core.data.dto.search import PaginatedElasticSearchDocumentHits
 from app.core.data.orm.util import get_parent_project_id
-from app.core.filters.column_info import ColumnInfo
-from app.core.filters.filtering import Filter
-from app.core.filters.sorting import Sort
+from app.core.search.column_info import ColumnInfo
+from app.core.search.filtering import Filter
 from app.core.search.memo_search.memo_search import memo_info, memo_search
 from app.core.search.memo_search.memo_search_columns import MemoColumns
+from app.core.search.sorting import Sort
 
 router = APIRouter(
     prefix="/memo", dependencies=[Depends(get_current_user)], tags=["memo"]
