+++ backend/src/app/core/analysis/timeline_analysis/timeline.py
@@ -16,11 +16,11 @@ from app.core.data.orm.source_document import SourceDocumentORM
 from app.core.data.orm.source_document_metadata import SourceDocumentMetadataORM
 from app.core.db.sql_service import SQLService
 from app.core.db.sql_utils import aggregate_ids
-from app.core.filters.column_info import (
+from app.core.search.column_info import (
     ColumnInfo,
 )
-from app.core.filters.filtering import Filter
-from app.core.filters.search_builder import SearchBuilder
+from app.core.search.filtering import Filter
+from app.core.search.search_builder import SearchBuilder
 
 
 def timeline_analysis_info(
