+++ backend/src/app/core/search/bbox_search/bbox_search.py
@@ -16,13 +16,13 @@ from app.core.data.orm.code import CodeORM
 from app.core.data.orm.source_document import SourceDocumentORM
 from app.core.data.repo.repo_service import RepoService
 from app.core.db.sql_service import SQLService
-from app.core.filters.column_info import (
+from app.core.search.bbox_search.bbox_search_columns import AnnotatedImagesColumns
+from app.core.search.column_info import (
     ColumnInfo,
 )
-from app.core.filters.filtering import Filter
-from app.core.filters.search_builder import SearchBuilder
-from app.core.filters.sorting import Sort
-from app.core.search.bbox_search.bbox_search_columns import AnnotatedImagesColumns
+from app.core.search.filtering import Filter
+from app.core.search.search_builder import SearchBuilder
+from app.core.search.sorting import Sort
 
 repo_service = RepoService()
