+++ backend/src/app/core/search/filtering_operators.py
