+++ backend/src/app/core/search/memo_search/memo_search.py
@@ -8,13 +8,13 @@ from app.core.data.orm.memo import MemoORM
 from app.core.data.orm.object_handle import ObjectHandleORM
 from app.core.db.elasticsearch_service import ElasticSearchService
 from app.core.db.sql_service import SQLService
-from app.core.filters.column_info import (
+from app.core.search.column_info import (
     ColumnInfo,
 )
-from app.core.filters.filtering import Filter
-from app.core.filters.search_builder import SearchBuilder
-from app.core.filters.sorting import Sort
+from app.core.search.filtering import Filter
 from app.core.search.memo_search.memo_search_columns import MemoColumns
+from app.core.search.search_builder import SearchBuilder
+from app.core.search.sorting import Sort
 
 
 def memo_info(
