+++ backend/src/app/core/search/memo_search/memo_search_columns.py
@@ -1,7 +1,7 @@
 from app.core.data.orm.memo import MemoORM
-from app.core.filters.column_info import AbstractColumns
-from app.core.filters.filtering_operators import FilterOperator, FilterValueType
-from app.core.filters.search_builder import SearchBuilder
+from app.core.search.column_info import AbstractColumns
+from app.core.search.filtering_operators import FilterOperator, FilterValueType
+from app.core.search.search_builder import SearchBuilder
 
 
 class MemoColumns(str, AbstractColumns):
