+++ backend/src/app/core/search/sdoc_search/sdoc_search.py
@@ -15,13 +15,13 @@ from app.core.data.orm.source_document import SourceDocumentORM
 from app.core.db.elasticsearch_service import ElasticSearchService
 from app.core.db.simsearch_service import SimSearchService
 from app.core.db.sql_service import SQLService
-from app.core.filters.column_info import ColumnInfo
-from app.core.filters.filtering import (
+from app.core.search.column_info import ColumnInfo
+from app.core.search.filtering import (
     Filter,
 )
-from app.core.filters.search_builder import SearchBuilder
-from app.core.filters.sorting import Sort
 from app.core.search.sdoc_search.sdoc_search_columns import SearchColumns
+from app.core.search.search_builder import SearchBuilder
+from app.core.search.sorting import Sort
 
 
 def search_info(project_id) -> List[ColumnInfo[SearchColumns]]:
