+++ backend/src/app/core/search/sorting.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 from sqlalchemy import asc, desc
 from sqlalchemy.orm import QueryableAttribute
 
-from app.core.filters.abstract_column import AbstractColumns
+from app.core.search.abstract_column import AbstractColumns
 
 T = TypeVar("T", bound=AbstractColumns)
